Remove all Dates & ajax, render just a 101x53 grid
This was an attempt to find the bottleneck in the rendering.

No real data is loaded from the server, and all date manipulation has
been bypassed. Even so, full page rendering takes about ten seconds.

It seems like just the way this app currently loops through to create
all the `Year` and `Week` components has a serious performance issue.

I'm not sure what to do about that.