Make adding dates work in not-Chrome
When you add an event, the date is now selected with a radio button,
instead of an `<input type="date">`. That means it works everywhere!

I also refactored `WeekDetail` heavily to make it easier to reason
about, splitting cohesive chunks of it into new components contained
within the same file.