Add basic sign up page. It sucks right now.
It doesn't even work yet.

But what does work? Linking to it! If you just go to "/signup", it
redirects to "/pricing". But if you go to "/signup/payment/annual", then
it displays.

What it displays, though. Ew. Not a way to get people to give me money!

So that's coming next. Cleaning up that page. And then making it
actually create an account!