Get sign up working
There's currently no checking on the uniqueness of the slug or email
address before the sign up is complete, at which point it just pops an
alert box. So that sucks! But otherwise it works. If everyone played by
the rules very well, no one would see that alert box. I'm not really
counting on that!