Fix bug that caused weird loading
It seemed like every now and then, things would load all wrong on the
calendar. Only a few dots would be considered "in the past" (and colored
like it). And every single dot said that it was for the current date
when hovered over.

I tracked this down to a race condition.

The `User` component fetches both the _User_, and the _Events_ for the
user. **Usually,** especially when there are a few events for a user,
the events come back second. That was the "expected" behavior (insofar
as I thought about this at all, previously). But when the events come
back first, then the `Life` component got rendered, and everything
looked stupid. This was compounded/enabled by the fact that I made the
`dateFor` function on `UserStore` return the current date, if the `born`
attribute wasn't yet set!

When a user first signs up, it seems like the call for their _Events_
was coming back first almost every single time. This led to a shitty
sign-up experience, since their calendar looked like non-functional
garbage upon first load.