Check if user still logged in on Google
This is a tricky thing to troubleshoot.

The user's Google session expires after about an hour. So we can't just
restore the session from their local storage and expect it to still be
accurate.

So, I check with the server. And if recording their login fails, I sign
them out.

This seemed to cause problems the one time that I was able to test it,
but I wasn't able to dig too deeply into what actually caused the
errors.