Add no-op help button; fix signin issue; add lib
The help button that will kick off the tour is now shown when you're on
your own calendar page.

The npm lib that will power the tour is installed.

A login bug is fixed. The redirect to your calendar was happening
unconditionally. That made sense when login only happened on one page.
But now that this is being fired everywhere, it needs to care about what
page you're on, and only redirect you if that makes sense.