Try using react-tourist for tour
Two problems with react-tourist

1. It puts the "look at this" bounding box _inside_ the element it tries
to highlight. And then tries to absolutely position it, assuming that
nothing up the DOM hierarchy is relatively or absolutely positioned.
Which, of course, things are in my app. So the pulsing "click here"
indicator is in a completely wrong spot.
2. It doesn't call `preventDefault` when the user clicks it. Since I
will be indicating that users should click _links_, navigation will
occur when I don't want it to.

I burned through three hours trying to troubleshoot and patch this
library. I have no idea how to troubleshoot installed npm modules! This
makes me very frustrated.