Fix sign up bug that caused calendar to not load
Their slug gets set while they're sitting there setting their birthdate.
Then when they get to the calendar page, that check that I had in there
was causing nothing to load.

We need to fetch the user and their events every time you load a user
page.

In order to make the saved app state still speed things up, I decided to
not flush the UserStore or EventStore state after a new user is
requested. So on reload, the calendar loads quickly from your
localStorage, then makes a request to the server, and will update if it
finds anything new.