Put Forest&Branches in own section; bigger font
The bigger font is a particularly interesting change.

Now for screens with a min width of 700px, the font size will be 1.5em.
Everywhere. Except two spots where I overrode back to the old value of
1.1rem.

I think this actually turned out to look pretty good. It gives the
desktop experience a little more of a mobile vibe. Everything feels
incredibly legible. I was able to make the future-dots the old
super-light color.